requirements.txt
Appium-Python-Client==2.11.1
astroid==2.15.6
attrs==23.1.0
black==23.7.0
certifi==2023.7.22
cffi==1.15.1
cfgv==3.4.0
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
dill==0.3.7
distlib==0.3.7
exceptiongroup==1.1.3
filelock==3.12.3
greenlet==2.0.2
h11==0.14.0
identify==2.5.27
idna==3.4
iniconfig==2.0.0
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
mypy-extensions==1.0.0
nodeenv==1.8.0
outcome==1.2.0
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
pluggy==1.2.0
pre-commit==3.3.3
pycparser==2.21
pyee==9.0.4
pylint==2.17.5
PySocks==1.7.1
pytest==7.4.0
pytest-base-url==2.0.0
python-slugify==8.0.1
PyYAML==6.0.1
requests==2.31.0
selenium==4.11.2
sniffio==1.3.0
sortedcontainers==2.4.0
text-unidecode==1.3
tomli==2.0.1
tomlkit==0.12.1
trio==0.22.2
trio-websocket==0.10.3
typing_extensions==4.7.1
urllib3==2.0.4
virtualenv==20.24.3
wrapt==1.15.0
wsproto==1.2.0
